Add files via upload
This program gives us the capability to multiple rate * pay including an overtime pay of 0.5 for any hours greater than 40. In this program i utilize Input, for entering the hours and rate of pay, float to uniform the numbers given, #print statements for notes to myself, if to run the calculation while determining if it is overtime pay, and else to calculate if it is regular pay.