Add files via upload
I utilize the last program example problem, but this time I have added a "try" and "except" to prevent tracebacks. In the case of a non-numeric input, the program would previously crash. after adding a try: and except: to print an error and end the program that is no longer the case. the program will now safely stop if given improper input.